drivers/radio/nrf24l01: Optimized status reading.
The value of the STATUS register is always transmitted by the
chip when reading any command. So a R_REGISTER command and the
turnaround time can be spared by issuing a NOP command instead.

This implementation suggested by the datasheet.

This operation is compatible with both nRF24L01 and nRF24L01+.

Signed-off-by: Marcell Pünkösd <punkosdmarcell@rocketmail.com>